Safari-611.1.10/LayoutTests/userscripts/user-script-video-document-expected.txt

CONSOLE MESSAGE: End Script: user-script-video-document.html
CONSOLE MESSAGE: Start Script: silence-loop.mov
CONSOLE MESSAGE: End Script: silence-loop.mov
